+++ src/getStepStartStates.test.ts
@@ -93,6 +93,7 @@ import {
   STEP_RM_EVENT_GRID_TOPICS,
   STEP_RM_EVENT_GRID_TOPIC_SUBSCRIPTIONS,
 } from './steps/resource-manager/event-grid';
+import { AdvisorSteps } from './steps/resource-manager/advisor';
 
 describe('getStepStartStates', () => {
   test('all steps represented', () => {
@@ -168,6 +169,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPIC_SUBSCRIPTIONS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPIC_SUBSCRIPTIONS]: { disabled: true },
+      [AdvisorSteps.RECOMMENDATIONS]: { disabled: true },
     });
   });
 
@@ -239,6 +241,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPIC_SUBSCRIPTIONS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPIC_SUBSCRIPTIONS]: { disabled: true },
+      [AdvisorSteps.RECOMMENDATIONS]: { disabled: true },
     });
   });
 
@@ -310,6 +313,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: false },
       [STEP_RM_EVENT_GRID_TOPIC_SUBSCRIPTIONS]: { disabled: false },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPIC_SUBSCRIPTIONS]: { disabled: false },
+      [AdvisorSteps.RECOMMENDATIONS]: { disabled: false },
     });
   });
 });