+++ src/index.ts
@@ -24,6 +24,7 @@ import { batchSteps } from './steps/resource-manager/batch';
 import { redisCacheSteps } from './steps/resource-manager/redis-cache';
 import { containerInstanceSteps } from './steps/resource-manager/container-instance';
 import { eventGridSteps } from './steps/resource-manager/event-grid';
+import { advisorSteps } from './steps/resource-manager/advisor';
 
 export function hasSubscriptionId(config: IntegrationConfig): boolean {
   const subscriptionId = config.subscriptionId;
@@ -81,5 +82,6 @@ export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = 
     // NOTE: Because any resource in Azure could be an Event Grid Topic, this step should be executed last. See SDK #326: https://github.com/JupiterOne/sdk/issues/326
     // This will ensure that other resources that an organization has can be tracked as 'topics' so that we can associate Event Grid Topic Subscriptions to them.
     ...eventGridSteps,
+    ...advisorSteps,
   ],
 };